+++ src/main/java/es/in2/issuer/backend/shared/domain/model/dto/credential/lear/machine/LEARCredentialMachine.java
@@ -45,7 +45,8 @@ public record LEARCredentialMachine(
                     @JsonProperty("organization") String organization,
                     @JsonProperty("country") String country,
                     @JsonProperty("commonName") String commonName,
-                    @JsonProperty("serialNumber") String serialNumber
+                    @JsonProperty("serialNumber") String serialNumber,
+                    @JsonProperty("email") String email
             ) {
             }