+++ src/main/java/es/in2/issuer/backend/shared/domain/util/Utils.java
@@ -42,9 +42,7 @@ public final class Utils {
     }
 
     public static LEARCredentialMachine.CredentialSubject.Mandate.Mandator extractMandatorLearCredentialMachine(LEARCredential credential) {
-        System.out.println("Credential: " + credential);
         List<String> types = credential.type();
-        System.out.println("Types: " + types);
         if (types.contains("LEARCredentialMachine")) {
             return ((LEARCredentialMachine) credential).credentialSubject().mandate().mandator();
         }