+++ src/main/java/es/in2/issuer/backend/shared/domain/util/factory/CredentialFactory.java
@@ -37,7 +37,7 @@ public class CredentialFactory {
                     .doOnSuccess(verifiableCertification -> log.info("ProcessID: {} - Label Credential mapped: {}", processId, credential));
         } else if(preSubmittedCredentialRequest.schema().equals(LEAR_CREDENTIAL_MACHINE)) {
             return learCredentialMachineFactory.mapAndBuildLEARCredentialMachine(credential, operationMode, email)
-                    .doOnSuccess(learCredentialEmployee -> log.info("ProcessID: {} - LEARCredentialEmployee mapped: {}", processId, credential));
+                    .doOnSuccess(learCredentialEmployee -> log.info("ProcessID: {} - LEARCredentialMachine mapped: {}", processId, credential));
         }
         return Mono.error(new CredentialTypeUnsupportedException(preSubmittedCredentialRequest.schema()));
     }
@@ -64,6 +64,9 @@ public class CredentialFactory {
             case LABEL_CREDENTIAL ->
                     labelCredentialFactory
                             .mapCredentialAndBindIssuerInToTheCredential(decodedCredential, procedureId);
+            case LEAR_CREDENTIAL_MACHINE ->
+                learCredentialMachineFactory
+                        .mapCredentialAndBindIssuerInToTheCredential(decodedCredential, procedureId);
             default ->
                     Mono.error(new CredentialTypeUnsupportedException(credentialType));
         };