+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [v2.3.0](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.3.0)
+### Added
+- Add LEARCredentialMachine issuance.
+- 
+## [v2.2.1](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.2.1)
+### Fixed
+- Revocation endpoint for label
 
 ## [v2.3.1](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.3.1)
 ### Fixed
@@ -19,7 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v2.1.2](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.1.2)
 ### Fixed
 - Fixed url backend in Credential Status
-
 
 ## [v2.1.1](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.1.1)
 ### Fixed